CR: update object structure
